+++ RELEASE_NOTES.md
@@ -36,6 +36,8 @@ been tested on Linux. As of this release, the latest known version to work with 
 
 ### New user-visible features
 
+- (internet) !2405 Added support for TCP FACK (Forward Acknowledgement).
+
 - (visualizer) Add Lr-Wpan NetDevices support to the Pyviz visualizer.
 
 ### Bugs fixed